lxd: move lxdclient from /provider to /tools
This is in preparation for adding the LXD container type.
