+++ provider/lxd/environ_broker.go
@@ -17,7 +17,7 @@ import (
 	"github.com/juju/juju/environs"
 	"github.com/juju/juju/instance"
 	"github.com/juju/juju/provider/common"
-	"github.com/juju/juju/provider/lxd/lxdclient"
+	"github.com/juju/juju/tools/lxdclient"
 	"github.com/juju/juju/state/multiwatcher"
 )
