+++ provider/lxd/instance_test.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/juju/juju/instance"
 	"github.com/juju/juju/provider/lxd"
-	"github.com/juju/juju/provider/lxd/lxdclient"
+	"github.com/juju/juju/tools/lxdclient"
 )
 
 type instanceSuite struct {
