+++ provider/lxd/lxd.go
@@ -9,7 +9,7 @@ import (
 	"github.com/juju/loggo"
 
 	"github.com/juju/juju/environs/tags"
-	"github.com/juju/juju/provider/lxd/lxdclient"
+	"github.com/juju/juju/tools/lxdclient"
 )
 
 // The metadata keys used when creating new instances.
