+++ tools/lxdclient/addserver.go
