+++ tools/lxdclient/addserver_test.go
