+++ tools/lxdclient/client_test.go
