+++ tools/lxdclient/config_test.go
