+++ tools/lxdclient/package_test.go
