+++ tools/lxdclient/remote_test.go
