+++ tools/lxdclient/utils.go
